Fix bug when restarting right after a round ends
When pressing Restart right after a round was
finished, the turn wouldn't swap (if swap every
round is on) and the gameDisplay was not updated
properly. The logic for restarting a game has now
been centralized into a single function. Restart
at this point will now do the same exact thing as
clicking on a cell.